Refractor paginator
