+++ src/AppBundle/Repository/CountryRepository.php
@@ -10,7 +10,7 @@ use Doctrine\ORM\Tools\Pagination\Paginator;
  * This class was generated by the Doctrine ORM. Add your own custom
  * repository methods below.
  */
-class CountryRepository extends \Doctrine\ORM\EntityRepository
+class CountryRepository extends BaseRepository
 {
     /**
     * Get all Countries with paginator and Order.
@@ -25,35 +25,6 @@ class CountryRepository extends \Doctrine\ORM\EntityRepository
                 ->getQuery();
 
         $paginator = $this->paginate($query, $currentPage, 10);
-
-        return $paginator;
-    }
-
-
-    /**
-     * Paginator Helper
-     *
-     * Pass through a query object, current page & limit
-     * the offset is calculated from the page and limit
-     * returns an `Paginator` instance, which you can call the following on:
-     *
-     *     $paginator->getIterator()->count() # Total fetched (ie: `5` posts)
-     *     $paginator->count() # Count of ALL posts (ie: `20` posts)
-     *     $paginator->getIterator() # ArrayIterator
-     *
-     * @param Doctrine\ORM\Query $dql   DQL Query Object
-     * @param integer            $page  Current page (defaults to 1)
-     * @param integer            $limit The total number per page (defaults to 5)
-     *
-     * @return \Doctrine\ORM\Tools\Pagination\Paginator
-     */
-    public function paginate($dql, $page = 1, $limit = 5)
-    {
-        $paginator = new Paginator($dql);
-
-        $paginator->getQuery()
-            ->setFirstResult($limit * ($page - 1)) // Offset
-            ->setMaxResults($limit); // Limit
 
         return $paginator;
     }
