+++ src/AppBundle/Repository/ImageRepository.php
@@ -8,6 +8,6 @@ namespace AppBundle\Repository;
  * This class was generated by the Doctrine ORM. Add your own custom
  * repository methods below.
  */
-class ImageRepository extends \Doctrine\ORM\EntityRepository
+class ImageRepository extends BaseRepository
 {
 }
